added BeerTableManager
