+++ HSE Punk Project/BeerTable/ViewController.swift
@@ -7,31 +7,15 @@
 
 import UIKit
 
-final class ViewController: UIViewController, UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return beerData.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let beer = beerData[indexPath.row]
-        var cell = UITableViewCell()
-        var configuration = cell.defaultContentConfiguration()
-//        configuration.image = UIImage()
-        configuration.text = beer.name
-        configuration.secondaryText = beer.tagline
-        cell.contentConfiguration = configuration
-        return cell
-    }
+final class ViewController: UIViewController {
 
     private lazy var tableView: UITableView = {
         let tableView = UITableView()
-        tableView.dataSource = self
+//        tableView.dataSource = self
         return tableView
     }()
 
     private lazy var spinnerView =  UIActivityIndicatorView(style: .large)
-
-    private var beerData: [BeerDTO] = []
 
     override func viewDidLoad() {
         super.viewDidLoad()
